requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.46
pymysql==1.0.3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]